i2c: designware: Always poll to lower interrupt overhead and save energy
On a laptop with an i9-13900H (Asus Zenbook UX6404), resting a finger on
the 140 Hz touchpad produces thousands of i2c interrupts per second,
thereby preventing the CPU from entering any package-level C-states. As a
result, the lost package C-state time causes the package power to skyrocket
to 8 W, from a mere 0.5 W before touching the touchpad.

By busy polling for i2c transaction completion instead of using an
interrupt, power consumption during the above scenario improves
significantly, with package power measuring at 3 W for a total of 5 W of
savings.

Therefore, always poll for better energy efficiency.

Note that on AMD it is better to keep the usleep_range() in the poll loop
in i2c_dw_wait_transfer(), whereas on Intel it is more energy efficient to
omit it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>